update: created TimeSlotView which is used in Calendar
