+++ frontend/src/app/components/Calendar.tsx
@@ -8,11 +8,10 @@ import Typography from "@mui/material/Typography";
 import Box from "@mui/material/Box";
 import Dialog from "@mui/material/Dialog";
 import GroomingFinalizationViewCard from "@/app/components/cards/GroomingFinalizationViewCard";
-import Button from "@mui/material/Button";
 import SubmitButton from "@/app/components/buttons/SubmitButton";
 import DialogTitle from "@mui/material/DialogTitle";
 import DialogContent from "@mui/material/DialogContent";
-import {Cancel} from "@mui/icons-material";
+import TimeSlotView from "@/app/components/TimeSlotView";
 
 interface CalendarProps {
     dog: Dog;
@@ -22,44 +21,9 @@ interface CalendarProps {
 export default function Calendar({dog, grooming}: CalendarProps) {
     const email = getEmail();
     const jwt = getJWT();
-    const [timeSlots, setTimeSlots] = useState<TimeSlot[] | null>(null);
     const [selectedTimeSlot, setSelectedTimeSlot] = useState<TimeSlot | null>(null);
-    const dayNames = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
     const [open, setOpen] = useState(false);
     const [response, setResponse] = useState<string>("");
-
-    const fetchTimeStamps = async () => {
-        if (dog && jwt && email && grooming)
-            try {
-                const response = await fetch(`/api/reservation/timeslots?email=${email}&dogName=${dog.name}&isShowerOnly=${grooming.showerOnly}`, {
-                    method: 'GET',
-                    headers: {
-                        Authorization: `Bearer ${jwt}`,
-                    },
-                });
-
-                if (response.ok) {
-                    const timeSlotData: TimeSlot[] = await response.json();
-                    setTimeSlots(timeSlotData);
-                }
-            } catch (error) {
-                console.error('Error fetching user data:', error);
-            }
-
-    }
-
-
-    function groupByDay(timeSlots: TimeSlot[]) {
-        return timeSlots.reduce((groups, timeSlot) => {
-            const date = new Date(timeSlot.start).getDay();
-            const dayName = dayNames[date];
-            if (!groups[dayName]) {
-                groups[dayName] = [];
-            }
-            groups[dayName].push(timeSlot);
-            return groups;
-        }, {} as Record<string, TimeSlot[]>);
-    }
 
     const handleAccept = () => async () => {
         try {
@@ -80,7 +44,6 @@ export default function Calendar({dog, grooming}: CalendarProps) {
                     setResponse("Reservation successful!");
                 } else {
                     setResponse("Reservation failed! Try again.");
-                    fetchTimeStamps();
                 }
             }
         } catch (error) {
@@ -93,34 +56,15 @@ export default function Calendar({dog, grooming}: CalendarProps) {
         setOpen(true);
     }
 
-    useEffect(() => {
-        openDialog();
-    }, [selectedTimeSlot])
+    const onTimeSlotSelect = (timeSlot: TimeSlot) => {
+        setSelectedTimeSlot(timeSlot);
+        openDialog()
+    }
 
-    useEffect(() => {
-        fetchTimeStamps();
-    }, [])
-
-    const groupedTimeSlots = groupByDay(timeSlots || []);
 
     return (
         <>
-            <Grid container spacing={2}>
-                {Object.entries(groupedTimeSlots).map(([day, timeSlots]) => (
-                    <Grid item xs={12} sm={6} md={4} lg={3} xl={2} key={day}>
-                        <Box display="flex" justifyContent="center" alignItems="center" height="100%">
-                            <Typography variant="h6">{day}</Typography>
-                        </Box>
-                        <Grid container direction="column" justifyContent="center" alignItems="center">
-                            {timeSlots.map((timeSlot, index) => (
-                                <Grid item key={index}>
-                                    <TimeSlotButton timeSlot={timeSlot} setSelectedTimeSlot={setSelectedTimeSlot}/>
-                                </Grid>
-                            ))}
-                        </Grid>
-                    </Grid>
-                ))}
-            </Grid>
+            <TimeSlotView dog={dog} grooming={grooming} onTimeSlotSelect={onTimeSlotSelect}/>
             <Dialog open={open}>
                 <DialogTitle>Finalize Your Booking</DialogTitle>
                 <DialogContent>
