+++ frontend/src/app/components/buttons/TimeSlotButton.tsx
@@ -3,17 +3,17 @@ import Button from "@mui/material/Button";
 
 interface TimeSlotProps {
     timeSlot: TimeSlot;
-    setSelectedTimeSlot: (timeSlot: TimeSlot) => void;
+    onTimeSlotSelect: (timeSlot: TimeSlot) => void;
 }
 
-export default function TimeSlotButton({timeSlot, setSelectedTimeSlot}: TimeSlotProps) {
+export default function TimeSlotButton({timeSlot, onTimeSlotSelect}: TimeSlotProps) {
     const timeSlotString = new Date(timeSlot.start).toLocaleTimeString([], {
         hour: '2-digit',
         minute: '2-digit',
         hour12: false
     });
     const handleClick = () => {
-        setSelectedTimeSlot(timeSlot);
+        onTimeSlotSelect(timeSlot);
     }
     return (
         <>
